Move PART handling into the messaging loop.
This makes sure we can part after clearing the message queues and wont
deliver messages after we have sent back a PART confirmation to the
client.

Hack around gamja reopening buffers if it gets messages after sending
the client PART which will usually happen when connected to a few
channels or always when connected to 1 channel.